# projects/datasite/internal/oas/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn8AllowedHeaders = map[string]string{
		"POST": "Authorization,Content-Type",
	}
	rn1AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn5AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PATCH":  "Authorization,Content-Type",
		"PUT":    "Authorization,Content-Type",
	}
	rn3AllowedHeaders = map[string]string{
		"GET":  "Authorization",
		"POST": "Authorization,Content-Type",
	}
	rn7AllowedHeaders = map[string]string{
		"DELETE": "Authorization",
		"GET":    "Authorization",
		"PATCH":  "Authorization,Content-Type",
		"PUT":    "Authorization,Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "admin/query"

				if l := len("admin/query"); len(elem) >= l && elem[0:l] == "admin/query" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleQuerySqlRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "POST",
							allowedHeaders: rn8AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}

			case 'c': // Prefix: "cinemas"

				if l := len("cinemas"); len(elem) >= l && elem[0:l] == "cinemas" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListCinemasRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleCreateCinemaRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET,POST",
							allowedHeaders: rn1AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleDeleteCinemaRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleGetCinemaRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PATCH":
							s.handlePatchCinemaRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PUT":
							s.handleUpdateCinemaRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,GET,PATCH,PUT",
								allowedHeaders: rn5AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "application/json",
							})
						}

						return
					}

				}

			case 'm': // Prefix: "movie_logs"

				if l := len("movie_logs"); len(elem) >= l && elem[0:l] == "movie_logs" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleListMovieLogsRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleCreateMovieLogRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET,POST",
							allowedHeaders: rn3AllowedHeaders,
							acceptPost:     "application/json",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleDeleteMovieLogRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleGetMovieLogRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PATCH":
							s.handlePatchMovieLogRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PUT":
							s.handleUpdateMovieLogRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "DELETE,GET,PATCH,PUT",
								allowedHeaders: rn7AllowedHeaders,
								acceptPost:     "",
								acceptPatch:    "application/json",
							})
						}

						return
					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "admin/query"

				if l := len("admin/query"); len(elem) >= l && elem[0:l] == "admin/query" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = QuerySqlOperation
						r.summary = "Execute a SQL query and return results"
						r.operationID = "querySql"
						r.operationGroup = ""
						r.pathPattern = "/admin/query"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'c': // Prefix: "cinemas"

				if l := len("cinemas"); len(elem) >= l && elem[0:l] == "cinemas" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListCinemasOperation
						r.summary = "List all cinemas"
						r.operationID = "listCinemas"
						r.operationGroup = ""
						r.pathPattern = "/cinemas"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = CreateCinemaOperation
						r.summary = "Create a cinema"
						r.operationID = "createCinema"
						r.operationGroup = ""
						r.pathPattern = "/cinemas"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = DeleteCinemaOperation
							r.summary = "Delete a cinema"
							r.operationID = "deleteCinema"
							r.operationGroup = ""
							r.pathPattern = "/cinemas/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = GetCinemaOperation
							r.summary = "Get a cinema by ID"
							r.operationID = "getCinema"
							r.operationGroup = ""
							r.pathPattern = "/cinemas/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PATCH":
							r.name = PatchCinemaOperation
							r.summary = "Partially update a cinema"
							r.operationID = "patchCinema"
							r.operationGroup = ""
							r.pathPattern = "/cinemas/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PUT":
							r.name = UpdateCinemaOperation
							r.summary = "Replace a cinema"
							r.operationID = "updateCinema"
							r.operationGroup = ""
							r.pathPattern = "/cinemas/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			case 'm': // Prefix: "movie_logs"

				if l := len("movie_logs"); len(elem) >= l && elem[0:l] == "movie_logs" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = ListMovieLogsOperation
						r.summary = "List all movie logs"
						r.operationID = "listMovieLogs"
						r.operationGroup = ""
						r.pathPattern = "/movie_logs"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = CreateMovieLogOperation
						r.summary = "Create a movie log"
						r.operationID = "createMovieLog"
						r.operationGroup = ""
						r.pathPattern = "/movie_logs"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "id"
					// Leaf parameter, slashes are prohibited
					idx := strings.IndexByte(elem, '/')
					if idx >= 0 {
						break
					}
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "DELETE":
							r.name = DeleteMovieLogOperation
							r.summary = "Delete a movie log"
							r.operationID = "deleteMovieLog"
							r.operationGroup = ""
							r.pathPattern = "/movie_logs/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = GetMovieLogOperation
							r.summary = "Get a movie log by ID"
							r.operationID = "getMovieLog"
							r.operationGroup = ""
							r.pathPattern = "/movie_logs/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PATCH":
							r.name = PatchMovieLogOperation
							r.summary = "Partially update a movie log"
							r.operationID = "patchMovieLog"
							r.operationGroup = ""
							r.pathPattern = "/movie_logs/{id}"
							r.args = args
							r.count = 1
							return r, true
						case "PUT":
							r.name = UpdateMovieLogOperation
							r.summary = "Replace a movie log"
							r.operationID = "updateMovieLog"
							r.operationGroup = ""
							r.pathPattern = "/movie_logs/{id}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}

				}

			}

		}
	}
	return r, false
}